Added casting to frontend

+++ src/main.rs
@@ -31,8 +31,8 @@ fn main() -> std::io::Result<()> {
     fn add(a: 10000xi64, b: 10000xi64, result: 10000xi64) i64 {
 	
         let mut m: 2xi64 = @{0_i64, 2_i64};
-    
-        @set.i64(c: m, i: 0_ii, v: 1000_i64);
+        
+        @set.i64(c: m, i: 0_ii, v: 1000_i32 as i64);
         @index.i64(c: m, i: 0_ii)
     };
